rename deploy to deployer
