+++ src/App/Commands/DeployCommand.php
@@ -11,9 +11,9 @@ use Console\App\Commands\Apps\AppsListCommand;
 use Console\App\Commands\Apps\AppsNewCommand;
 use Console\App\Commands\Databases\DatabasesDescribeCommand;
 use Console\App\Commands\Databases\DatabasesListCommand;
-use Console\App\Deploy\DeployInterface;
-use Console\App\Deploy\Laravel;
-use Console\App\Deploy\Symfony;
+use Console\App\Deployers\DeployInterface;
+use Console\App\Deployers\Laravel;
+use Console\App\Deployers\Symfony;
 use Console\App\Helpers\ConfigHelper;
 use Console\App\Helpers\DeployHelper;
 use GuzzleHttp\ClientInterface;
